[dev] 1.fix printf %d , when %d = 0 problem 2.fix new line problem

+++ 04-TimerInterrupt/riscv.c
@@ -2,7 +2,7 @@
 
 
 
-uint8_t r_mhartid(void){
+uint32_t r_mhartid(void){
     reg_t x;
     asm volatile( "csrr %0, mhartid" : "=r" (x) );
     return x;
